services/orchestrator/requirements.txt
fastapi
uvicorn
httpx
pydantic
google-genai
langgraph
chromadb